Quarterly mean arrival delay times for a set of 402 flights

Created from nycflights13::flights

Usage

quarterly_flights

Format

A data frame with 1608 rows and 6 variables

tailnum

a unique identifier created from tailnum, origin, destination and carrier

carrier

carrier code

origin

origin code

dest

destination code

qu

quarter

mean_arr_delay

average delay on arrival as either on_time or late

Source

nycflights13::flights