tidy up dataframe containing model feature importance

returns dataframe with exactly two columns, vars and imp and aggregates dummy encoded variables. Helper function called by all functions that take an imp parameter. Can be called manually if formula for aggregating dummy encoded variables must be modified.

tidy_imp(imp, df, .f = max)

Arguments

imp
dataframe or matrix with feature importance information

df
dataframe, modelling training data

.f
window function, Default: max

Value

dataframe

vars

character column with feature names

imp

numerical column, importance values

Examples

# randomforest
df = mtcars2[, ! names(mtcars2) %in% 'ids' ]
m = randomForest::randomForest( disp ~ ., df)
imp = m$importance
tidy_imp(imp, mtcars2)

#> # A tibble: 10 x 2
#>    vars      imp
#>    <chr>   <dbl>
#>  1 cyl   104496.
#>  2 mpg    79654.
#>  3 hp     78618.
#>  4 wt     69626.
#>  5 drat   49637.
#>  6 gear   23535.
#>  7 qsec   20962.
#>  8 vs     14347.
#>  9 carb    8955.
#> 10 am      3045.